backend/requirements.txt
fastapi
uvicorn[standard]
pillow
numpy
deepface
transformers
torch
requests
opencv-python-headless
